+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## v0.10.0 (TBD)
 
+### Enhancements
+
+- Added support for gRPC server side reflection to all components (#949).
+
 ### Changes
 
 - Separated the store API into three separate services (#932).
+- Added a faucet Dockerfile (#933).
 
 ## v0.9.0 (2025-05-30)
 
@@ -16,6 +21,7 @@
 - Introduced Network Transaction Builder (#840).
 - Added way of executing and proving network transactions (#841).
 - [BREAKING] Add HTTP ACCEPT header layer to RPC server to enforce semver requirements against client connections (#844).
+- Add miden-proving-service and miden-proving-service-client crates (#926).
 
 ### Changes